pyproject.toml
[project]
name = "nonebot-plugin-terralink"
version = "0.3.2"
description = "实现泰拉瑞亚 TModLoader 服务器与 QQ 群双向互通的 NoneBot2 插件"
authors = [
    { name = "newcovid", email = "280310454@qq.com" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "GPL-3.0" }
keywords = ["nonebot", "terraria", "tmodloader", "bridge", "server", "websocket"]
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-adapter-onebot>=2.3.0",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "nonebot-plugin-htmlrender>=0.4.0"
]

[project.urls]
Homepage = "https://github.com/newcovid/nonebot-plugin-terralink"
Documentation = "https://github.com/newcovid/nonebot-plugin-terralink#readme"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

# [修复] 使用 find 指令自动查找子包，防止 core/services/matchers 丢失
[tool.setuptools.packages.find]
where = ["."]
include = ["nonebot_plugin_terralink", "nonebot_plugin_terralink.*"]